Break down SCIM modules and annotate SCIM tests (#614)
